# added numba to ewma and a few improvs