New Feature

Added numba to ewma from https://stackoverflow.com/a/51392341/4013571
Also included the infinite history EWMA which is **much** faster.

Personally I prefer using span / window rather than halflife as it is more intuitive but I converted the `window` to `halflife` to match the existing repo style.

Lmk if this doesn't fit

- [x ] Code conforms to style guide
- [ x] Doctests
- [ ] Unit-tests
- [x ] Documentation
- [] New line item in [releases.md](https://github.com/fastats/fastats/blob/master/releases.md)
- [ ] Info added to `literature` directory

Bug Report

- [x ] Minimal repro in a test function