@dwillmer 2019-02-06
Hi @flipdazed,

Firstly thanks very much for taking the time to open a PR - developer free time is a scarce resource, and we're always grateful when people join in (and always looking for people who enjoy numerical coding! :smile:).

However, there's a few issues with this PR that you've probably noticed from the failing tests (our tests generally don't fail without good reason, it isn't that sort of project :smile:).

To start with, I think you've misunderstood part of the reasoning behind the fastats library - you'll notice from most of the rest of the repo that we don't *tend* to put the numba decorator on our maths functions, even though we use numba for pretty much everything. This is because the numba decorator is automatically applied by the framework when using, for example, the `windowed_pass` function. All of the CI jobs failed, and almost all of the failures were due to this.

If a user wants to jit the function themselves, that's fine, but we choose not to do it automatically for a number of reasons; for example by **not** applying the decorator we can run simple unit-tests on both the python and the jitted functions to ensure the outputs are the same within bounds. 

You will see a variety of unit-tests in our `tests` folder which run parameterised or generated tests over both the raw python function and the jitted function, and compare with either hardcoded outputs
or the pandas/statsmodels equivalent. These tests have caught issues in the past, and it's a pattern we quite like, so I'd rather not introduce the `jit` decorator in this file. There are places where we decorate the individual functions, but those are generally exceptions.

Secondly, you've rearranged the algorithm without making any significant changes to it, which doesn't really have any benefit:
<img width="570" alt="screen shot 2019-02-06 at 20 45 07" src="https://user-images.githubusercontent.com/2343358/52372462-22234d80-2a50-11e9-9217-5f8715f1a3af.png">

(the first one is the existing function, the second is yours)

However you have removed the assertion, which is very important for a number of reasons and is the cause of one of the failing tests in each of the CI runs. The assertion is extremely important for numba - the numba library (read: LLVM) can apply further optimizations after an assertion because it can do things such as removing bounds checks. In this case it's not used for looping/bounds checking, but the principle remains the same. As a word of advice, you shouldn't remove other people's assertions from the code - if a developer thinks it's important to blow up due to invalid input, that's something to be aware of.

Thirdly, you've not only removed one of the unit-tests with hard-coded expected values (useful for sanity checking), but you've replaced it with a check that requires pandas, and so it will be less clear if something has changed in the future. I personally like to have a mixture of hard-coded expected values and comparisons with other library functions - the more checks the better!

Fourthly, I appreciate you've called it the 'infinite_history' function, and put in a message about the accuracy, but really the return values are so incorrect in such a short series of iterations that i see no use for this algorithm at all.

Please don't feel like this is bad feedback - I really do appreciate the time you've put into this; and personally I think the comments you make above are the start of a really useful discussion about improving this part of the codebase. I have a few ideas that could improve the current state, that might be useful to code up:

- Add an API to allow the use of `span`/`window` rather than halflife (exactly as you say above - ideally this would be a wrapper function which calls into the same underlying function we have now).
- Add your doctest to the existing one (rather than removing the existing one) and expand on the existing EWMA tests.
- Expand on the existing assertion in order to encourage numba/LLVM to allow further optimizations

There is one unit-test set to `xfail` in our suite, and it is the one for proper nan handling in the EWMA function; I would normally recommend starting with this as a way into the codebase, however one member of my team has already completed this as part of a project at work and we're contemplating open-sourcing it, potentially in a different repository, so I'd recommend against taking this one on.

Thanks again for your time on this - I really do hope this encourages you to contribute in the future,

Cheers,
Dave