@flipdazed 2019-02-07
Hi no worries. There wasn't a contrib guide so wasn't entirely sure on where to go on tests and didn't want to spend more than a coffee break pushing!

I saw your EWMA was still in the todo list as were most of the other funcs so assumed you just hadn't decorated yet.