@AaronCritchley 2019-02-07
Hey,

Thanks for the feedback, there's a contribution guide in the README, I'll think about ways that we can make it clearer though!
(Link, for reference: https://github.com/fastats/fastats/blob/master/.github/CONTRIBUTING.md)

Good point about the TODO list, we should tidy that up, sorry that it was misleading in this case.

RE branching: I think the branching comes out with the LLVM optimisations in Numba but I'd have to look at the LLVM IR to confirm.

RE timing: For an array that small, your approach would likely be faster due to the lack of assertions at the start of the function (which are super helpful as Dave expanded on above), here are the results from benchmarking on a 1-D array locally:


```
Res for size: 10 ^ 3
Pandas: 729 µs ± 12.8 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
Old: 36.8 µs ± 1.59 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
New: 34.9 µs ± 746 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

Res for size: 10 ^ 4
Pandas: 1.26 ms ± 38.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
Old: 609 µs ± 5.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
New: 704 µs ± 2.71 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

Res for size: 10 ^ 5
Pandas: 7.52 ms ± 263 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Old: 6.68 ms ± 121 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
New: 7.33 ms ± 25.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

Res for size: 10 ^ 6
Pandas: 72.7 ms ± 1.69 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Old: 70.2 ms ± 236 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
New: 72.3 ms ± 1.48 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

Res for size: 10 ^ 7
Pandas: 720 ms ± 5.73 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
Old: 711 ms ± 12.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
New: 717 ms ± 10.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```