@AaronCritchley 2019-02-07
Pandas :   df.ewm(halflife=hl).mean()
Old       :  The original `ewma` implementation
New     :  The proposed `ewma` implementation

Where `size` is the number given to np.random.rand to generate the 1-d array (both array and dataframe creation are outside of the timing, for removal of doubt).

Benchmarking was done on:
CPU - Intel Core i7-6700 CPU @ 3.40GHz, 3408 Mhz, 4 Cores, 8 Logical Processors
RAM - 32GB 2400Hz DDR4
Numba - 0.42.0